Fixed: assertion on startup under wx29 Fixed: navigation dialog now showing under wx29 (linux) Fixed: Performance improved of the editor under wx29
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@4490 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ LiteEditor/frame.cpp
@@ -3185,7 +3185,7 @@ void clMainFrame::OnNewVersionAvailable(wxCommandEvent& e)
 
 		} else {
 			if (!data->GetShowMessage()) {
-				wxLogMessage(wxString::Format(wxT("Info: CodeLite is up-to-date (or newer), version used: %d, version on site:%d"), data->GetCurrentVersion(), data->GetNewVersion()));
+				wxLogMessage(wxString::Format(wxT("Info: CodeLite is up-to-date (or newer), version used: %d, version on site:%d"), (int)data->GetCurrentVersion(), (int)data->GetNewVersion()));
 
 			} else {
 				// User initiated the version check request
